+++ kinto/core/permission/__init__.py
@@ -181,7 +181,7 @@ class PermissionBase:
 
 def heartbeat(backend):
     def ping(request):
-        """Test the permission backend is operationnal.
+        """Test the permission backend is operational.
 
         :param request: current request object
         :type request: :class:`~pyramid:pyramid.request.Request`
@@ -195,7 +195,7 @@ def heartbeat(backend):
             else:
                 backend.add_user_principal(__HEARTBEAT_KEY__, 'alive')
                 backend.remove_user_principal(__HEARTBEAT_KEY__, 'alive')
-        except:
+        except Exception:
             logger.exception('Heartbeat Error')
             return False
         return True