+++ kinto/core/storage/postgresql/__init__.py
@@ -69,7 +69,7 @@ class Storage(StorageBase):
 
     """  # NOQA
 
-    schema_version = 15
+    schema_version = 17
 
     def __init__(self, client, max_fetch_size, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -206,15 +206,59 @@ class Storage(StorageBase):
         logger.debug('Flushed PostgreSQL storage tables')
 
     def collection_timestamp(self, collection_id, parent_id, auth=None):
-        query = """
-        SELECT as_epoch(collection_timestamp(:parent_id, :collection_id))
-            AS last_modified;
+        query_existing = """
+        WITH existing_timestamps AS (
+          -- Timestamp of latest record.
+          (
+            SELECT last_modified, as_epoch(last_modified) AS last_epoch
+            FROM records
+            WHERE parent_id = :parent_id
+              AND collection_id = :collection_id
+            ORDER BY last_modified DESC
+            LIMIT 1
+          )
+          -- Timestamp of latest tombstone.
+          UNION
+          (
+            SELECT last_modified, as_epoch(last_modified) AS last_epoch
+            FROM deleted
+            WHERE parent_id = :parent_id
+              AND collection_id = :collection_id
+            ORDER BY last_modified DESC
+            LIMIT 1
+          )
+          -- Timestamp of empty collection.
+          UNION
+          (
+            SELECT last_modified, as_epoch(last_modified) AS last_epoch
+            FROM timestamps
+            WHERE parent_id = :parent_id
+              AND collection_id = :collection_id
+          )
+        )
+        SELECT MAX(last_modified) AS last_modified, MAX(last_epoch) AS last_epoch
+          FROM existing_timestamps
         """
+
+        create_if_missing = """
+        INSERT INTO timestamps (parent_id, collection_id, last_modified)
+        VALUES (:parent_id, :collection_id, COALESCE(:last_modified, clock_timestamp()::timestamp))
+        ON CONFLICT (parent_id, collection_id) DO NOTHING
+        """
+
         placeholders = dict(parent_id=parent_id, collection_id=collection_id)
         with self.client.connect(readonly=False) as conn:
-            result = conn.execute(query, placeholders)
+            existing_ts = None
+            ts_result = conn.execute(query_existing, placeholders)
+            row = ts_result.fetchone()  # Will return (None, None) when empty.
+            existing_ts = row['last_modified']
+
+            conn.execute(create_if_missing, dict(last_modified=existing_ts, **placeholders))
+
+            result = conn.execute(query_existing, placeholders)
             record = result.fetchone()
-        return record['last_modified']
+
+        return record['last_epoch']
 
     def create(self, collection_id, parent_id, record, id_generator=None,
                id_field=DEFAULT_ID_FIELD,
@@ -865,14 +909,13 @@ class Storage(StorageBase):
             elif sort.field == modified_field:
                 sql_field = 'last_modified'
             else:
-                # Subfields: ``person.name`` becomes ``data->person->>name``
+                # Subfields: ``person.name`` becomes ``data->person->name``
                 subfields = sort.field.split('.')
                 sql_field = 'data'
                 for j, subfield in enumerate(subfields):
                     # Safely escape field name
                     field_holder = 'sort_field_{}_{}'.format(i, j)
                     holders[field_holder] = subfield
-                    # Use ->> to convert the last level to text.
                     sql_field += '->(:{})'.format(field_holder)
 
             sql_direction = 'ASC' if sort.direction > 0 else 'DESC'