+++ tests/core/test_logging.py
@@ -1,10 +1,11 @@
+import json
+import logging
 import unittest
 
 import mock
 from pyramid import testing
 
-from kinto.core import DEFAULT_SETTINGS
-from kinto.core import initialization
+from kinto.core import DEFAULT_SETTINGS, initialization, JsonLogFormatter
 
 from .support import BaseWebTest
 
@@ -25,22 +26,28 @@ class RequestSummaryTest(BaseWebTest, unittest.TestCase):
         return kwargs['extra']
 
     def test_standard_info_is_bound(self):
-        self.app.get('/', headers=self.headers)
+        headers = {
+            'User-Agent': 'Smith',
+            **self.headers
+        }
+        self.app.get('/', headers=headers)
         event_dict = self.logger_context()
         self.assertEqual(event_dict['path'], '/v0/')
         self.assertEqual(event_dict['method'], 'GET')
         self.assertEqual(event_dict['code'], 200)
+        self.assertEqual(event_dict['agent'], 'Smith')
         self.assertIsNotNone(event_dict['uid'])
         self.assertIsNotNone(event_dict['time'])
         self.assertIsNotNone(event_dict['t'])
-        self.assertIsNone(event_dict['agent'])
-        self.assertIsNone(event_dict['lang'])
-        self.assertIsNone(event_dict['errno'])
+        self.assertEqual(event_dict['errno'], 0)
+        self.assertNotIn('lang', event_dict)
+        self.assertNotIn('headers', event_dict)
+        self.assertNotIn('body', event_dict)
 
     def test_userid_is_none_when_anonymous(self):
         self.app.get('/')
         event_dict = self.logger_context()
-        self.assertIsNone(event_dict['uid'])
+        self.assertNotIn('uid', event_dict)
 
     def test_lang_is_not_none_when_provided(self):
         self.app.get('/', headers={'Accept-Language': 'fr-FR'})
@@ -62,6 +69,29 @@ class RequestSummaryTest(BaseWebTest, unittest.TestCase):
         app.get('/mushrooms', headers={'Authorization': 'Basic bWF0OjE='})
         event_dict = self.logger_context()
         self.assertEqual(event_dict['authn_type'], 'basicauth')
+
+    def test_headers_and_body_when_level_is_debug(self):
+        self.mocked.level = logging.DEBUG
+        body = b'{"boom": 1}'
+        self.app.post('/batch', body, headers=self.headers, status=400)
+        event_dict = self.logger_context()
+        self.assertEqual(event_dict['headers'], {
+            'Authorization': 'Basic bWF0OnNlY3JldA==',
+            'Content-Length': '11',
+            'Content-Type': 'application/json',
+            'Host': 'localhost:80'})
+        self.assertEqual(event_dict['body'], body)
+
+        self.maxDiff = None
+
+        responseBody = event_dict['response']['body']
+        self.assertEqual(json.loads(responseBody.decode('utf-8'))['error'], 'Invalid parameters')
+        responseHeaders = event_dict['response']['headers']
+        self.assertEqual(sorted(responseHeaders.keys()), [
+            'Access-Control-Expose-Headers',
+            'Content-Length',
+            'Content-Type',
+            'X-Content-Type-Options'])
 
 
 class BatchSubrequestTest(BaseWebTest, unittest.TestCase):
@@ -104,3 +134,22 @@ class BatchSubrequestTest(BaseWebTest, unittest.TestCase):
 
     def test_subrequests_are_logged_as_subrequest_summary(self):
         self.assertEqual(self.subrequest_mocked.info.call_count, 2)
+        args, kwargs = self.subrequest_mocked.info.call_args_list[-1]
+        extra = kwargs['extra']
+        self.assertEqual(extra['path'], '/v0/unknown2')
+        args, kwargs = self.subrequest_mocked.info.call_args_list[-2]
+        extra = kwargs['extra']
+        self.assertEqual(extra['path'], '/v0/unknown')
+
+
+class JsonFormatterTest(unittest.TestCase):
+    def test_logger_name(self):
+        JsonLogFormatter.init_from_settings({'project_name': 'kintowe'})
+        f = JsonLogFormatter()
+        record = logging.LogRecord('app.log', logging.DEBUG, '', 0, 'coucou', (), None)
+        result = f.format(record)
+        logged = json.loads(result)
+        self.assertEqual(logged['Logger'], 'kintowe')
+        self.assertEqual(logged['Type'], 'app.log')
+        # See https://github.com/mozilla/mozilla-cloud-services-logger/issues/2
+        self.assertEqual(logged['Fields']['msg'], 'coucou')